chore: add both version of readme
If you plan to publish your theme on WordPress Theme directory, the
readme.txt file can be useful. I'm not sure this is required for themes
(it is for plugins), but it is at least recommended.
See: https://make.wordpress.org/themes/2015/04/29/a-revised-readme/
The README.md allows better styling to publish the theme on a VCS.