feat: use rollup and dotenv to bundle JS & minify or add sourcemaps
JS is now using rollup to bundle all JS files with an entry point.
So it is no longer needed to concat a list of files. We can import
the desired components/vendors files in an entry point then rollup
bundle it!
Concerning minify/sourcemaps: same as CSS rewrite, I think
sourcemaps are useless in prod and we don't need minify version in
dev mode.