feat: add a script to take a screenshot of your homepage
I choose to use playwright-firefox for this task because it is smaller
than the chrome version and the webkit version does not work for me.
I also moved the config from gulp folder to its parent since I am
using some data from `.env`. And since BrowserSync is not the only
one using hostname and protocol, I reorganized a bit the
`.env.example` file.